fix(types): annotate 11 trivial modules (annotation-only)
Resolve the long tail of single-error mypy strict findings so these
modules stay fully type-checked and never enter the baseline:

- request_metrics: annotate call_next (RequestResponseEndpoint)
- validators / metal_inventory: annotate field_validator info/v params
- time_entry_metadata: Iterator[str] return on the string-walk generator
- material: annotate from_material classmethod
- customer / idempotency: dict[str, Any] type params
- admin_scan_metrics: float() the percentile return (no-any-return)
- consultation: type: ignore[assignment] for the PATCH Optional-widening
- repair_photo_service / consultation_photo_service: cast(str) the ORM
  Column[str] file_path (zero-runtime, matches existing repo pattern)

No behavior change.

+++ src/goldsmith_erp/api/routers/admin_scan_metrics.py
@@ -302,7 +302,7 @@ async def _fab_tap_latency(
     def _pct(p: float) -> float:
         # Nearest-rank percentile; good enough for small samples.
         idx = max(0, min(len(deltas_ms) - 1, int(round(p * (len(deltas_ms) - 1)))))
-        return round(deltas_ms[idx], 1)
+        return float(round(deltas_ms[idx], 1))
 
     return _pct(0.5), _pct(0.95)
